Updating to enable PII - Azure Table Storage Client side encryption Moved to using CloudTable bindings instead of IQueryable<isv> where encryption is required as it's easier to pass the encryptionoptions in this way. Added KeyVault to ARM template + Placeholder Settings in app service. Added some KeyVault key locator code.
Refactored code to not include a common DAL dependency on the KeyVault libraries so we can import them in the script format
Removed SharedData.csx as it is now no longer needed.
Updated readme to note that you now have to create and authorise the key to the app via an azure ad secret, and add access permissions in keyvault.
Updated configure-local-debug.ps1 to bring in new keyvault settings as environment variables.